+++ src/components/Welcome.js
@@ -1,23 +1,13 @@
 import React from "react";
-import GoogleSignin from "../img/btn_google_signin_dark_pressed_web.png";
+import { AuthButton } from "./AuthButton";
 
 const Welcome = () => {
-  const googleSignIn = () => {
-  };
-
   return (
     <main className="welcome">
       <h2>Welcome to React Chat.</h2>
       <img src="/logo512.png" alt="ReactJs logo" width={50} height={50} />
       <p>Sign in with Google to chat with with your fellow React Developers.</p>
-      <button className="sign-in">
-        <img
-          onClick={googleSignIn}
-          src={GoogleSignin}
-          alt="sign in with google"
-          type="button"
-        />
-      </button>
+      <AuthButton/>
     </main>
   );
 };
